Add "fit_always" option for supersized
Oddly, the current default for supersized is to scale up the image to fit the
width of the screen, which means that an image sized at 960x720 for a
1280x720 screen will be scaled up to fit the width and have the head and
tail chopped off.

We generally want to display the whole image, even if it means some
black bars, so it turns out that the "fit_always" option for supersized
does just that. From
http://buildinternet.com/project/supersized/docs.html#options_docs:

fit_always

Prevents the image from ever being cropped. Ignores minimum width and
height.

MAKE IT SO.

Signed-off-by: Dan Scott <dscott@laurentian.ca>